+++ FlatFiles/ByteColumn.cs
@@ -40,8 +40,7 @@ namespace FlatFiles
         /// <returns>The parsed byte value.</returns>
         protected override byte OnParse(IColumnContext context, string value)
         {
-            IFormatProvider provider = FormatProvider ?? CultureInfo.CurrentCulture;
-            return Byte.Parse(value, NumberStyles, provider);
+            return Byte.Parse(value, NumberStyles, GetFormatProvider(context, FormatProvider));
         }
 
         /// <summary>
@@ -52,11 +51,12 @@ namespace FlatFiles
         /// <returns>The formatted value.</returns>
         protected override string OnFormat(IColumnContext context, byte value)
         {
+            var provider = GetFormatProvider(context, FormatProvider);
             if (OutputFormat == null)
             {
-                return value.ToString(FormatProvider ?? CultureInfo.CurrentCulture);
+                return value.ToString(provider);
             }
-            return value.ToString(OutputFormat, FormatProvider ?? CultureInfo.CurrentCulture);
+            return value.ToString(OutputFormat, provider);
         }
     }
 }