+++ FlatFiles/IOptions.cs
@@ -1,4 +1,6 @@
-﻿namespace FlatFiles
+﻿using System;
+
+namespace FlatFiles
 {
     /// <summary>
     /// Represents reader/writer options that are common among file types.
@@ -14,5 +16,10 @@
         /// Gets whether column-level metadata should be disabled for non-metadata columns.
         /// </summary>
         bool IsColumnContextDisabled { get; }
+
+        /// <summary>
+        /// Gets the global, default format provider to use.
+        /// </summary>
+        IFormatProvider FormatProvider { get; }
     }
 }